Renamed config file into constants.js
